IGNITE-19212 Re-factoring
