+++ modules/platforms/cpp/tests/odbc-test/meta_queries_test.cpp
@@ -172,18 +172,18 @@ void check_column_meta_with_sqlcol_attribute(SQLHSTMT stmt, SQLUSMALLINT idx, co
 class meta_queries_test : public odbc_suite {
 public:
     static void SetUpTestSuite() {
-        odbc_connection conn(get_basic_connection_string());
-        conn.connect();
+        odbc_connection conn;
+        conn.odbc_connect(get_basic_connection_string());
 
-        SQLRETURN ret = conn.exec_query("CREATE TABLE META_QUERIES_TEST(ID INT PRIMARY KEY, VAL VARCHAR(60))");
+        SQLRETURN ret = conn.exec_query("CREATE TABLE META_QUERIES_TEST(ID INT PRIMARY KEY, STR VARCHAR(60))");
         if (!SQL_SUCCEEDED(ret)) {
             FAIL() << conn.get_statement_error_message();
         }
     }
 
     static void TearDownTestSuite() {
-        odbc_connection conn(get_basic_connection_string());
-        conn.connect();
+        odbc_connection conn;
+        conn.odbc_connect(get_basic_connection_string());
 
         conn.exec_query("DROP TABLE META_QUERIES_TEST");
     }
@@ -446,7 +446,7 @@ TEST_F(meta_queries_test, test_col_attributes_column_length)
 {
     odbc_connect(get_basic_connection_string());
 
-    SQLCHAR req[] = "select str from TBL_ALL_COLUMNS_SQL";
+    SQLCHAR req[] = "select str from META_QUERIES_TEST";
     SQLExecDirect(m_statement, req, SQL_NTS);
 
     SQLLEN int_val;
@@ -465,7 +465,7 @@ TEST_F(meta_queries_test, test_col_attributes_column_presicion)
 {
     odbc_connect(get_basic_connection_string());
 
-    SQLCHAR req[] = "select str from TBL_ALL_COLUMNS_SQL";
+    SQLCHAR req[] = "select str from META_QUERIES_TEST";
     SQLExecDirect(m_statement, req, SQL_NTS);
 
     SQLLEN int_val;
@@ -497,6 +497,8 @@ TEST_F(meta_queries_test, test_col_attributes_column_scale)
         FAIL() << (get_odbc_error_message(SQL_HANDLE_STMT, m_statement));
 }
 
+// TODO: IGNITE-19854 Implement metadata fetching for the non-executed query.
+#ifdef MUTED
 TEST_F(meta_queries_test, test_col_attributes_column_length_prepare)
 {
     odbc_connect(get_basic_connection_string());
@@ -585,7 +587,10 @@ TEST_F(meta_queries_test, test_col_attributes_column_scale_prepare)
     if (!SQL_SUCCEEDED(ret))
         FAIL() << (get_odbc_error_message(SQL_HANDLE_STMT, m_statement));
 }
+#endif //MUTED
 
+// TODO: IGNITE-19216 Implement type info query.
+#ifdef MUTED
 TEST_F(meta_queries_test, test_get_data_with_get_type_info)
 {
     odbc_connect(get_basic_connection_string());
@@ -597,7 +602,10 @@ TEST_F(meta_queries_test, test_get_data_with_get_type_info)
 
     check_single_row_result_set_with_get_data(m_statement);
 }
+#endif //MUTED
 
+// TODO: IGNITE-19214 Implement tables metadata fetching
+#ifdef MUTED
 TEST_F(meta_queries_test, test_get_data_with_tables)
 {
     odbc_connect(get_basic_connection_string());
@@ -612,7 +620,10 @@ TEST_F(meta_queries_test, test_get_data_with_tables)
 
     check_single_row_result_set_with_get_data(m_statement);
 }
+#endif //MUTED
 
+// TODO: IGNITE-19214 Implement table column metadata fetching
+#ifdef MUTED
 TEST_F(meta_queries_test, test_get_data_with_columns)
 {
     odbc_connect(get_basic_connection_string());
@@ -628,24 +639,26 @@ TEST_F(meta_queries_test, test_get_data_with_columns)
 
     check_single_row_result_set_with_get_data(m_statement);
 }
+#endif //MUTED
 
 TEST_F(meta_queries_test, test_get_data_with_select_query)
 {
     odbc_connect(get_basic_connection_string());
 
-    SQLCHAR insert_req[] = "insert into TestType(_key, str) VALUES(1, 'Lorem ipsum')";
+    SQLCHAR insert_req[] = "insert into META_QUERIES_TEST(id, str) VALUES(1, 'Lorem ipsum')";
     SQLRETURN ret = SQLExecDirect(m_statement, insert_req, SQL_NTS);
 
     if (!SQL_SUCCEEDED(ret))
         FAIL() << (get_odbc_error_message(SQL_HANDLE_STMT, m_statement));
 
-    SQLCHAR select_req[] = "select str from TBL_ALL_COLUMNS_SQL";
+    SQLCHAR select_req[] = "select str from META_QUERIES_TEST";
     ret = SQLExecDirect(m_statement, select_req, SQL_NTS);
 
     if (!SQL_SUCCEEDED(ret))
         FAIL() << (get_odbc_error_message(SQL_HANDLE_STMT, m_statement));
 
-    check_single_row_result_set_with_get_data(m_statement);
+    // TODO: IGNITE-19213 Implement data fetching
+//    check_single_row_result_set_with_get_data(m_statement);
 }
 
 TEST_F(meta_queries_test, test_insert_too_long_value_fail)
@@ -653,7 +666,8 @@ TEST_F(meta_queries_test, test_insert_too_long_value_fail)
     odbc_connect(get_basic_connection_string());
 
     SQLCHAR insert_req[] =
-        "insert into TestType(_key, str) VALUES(42, '0123456789012345678901234567890123456789012345678901234567891')";
+        "insert into META_QUERIES_TEST(id, str) "
+        "VALUES(42, '0123456789012345678901234567890123456789012345678901234567891')";
 
     SQLRETURN ret = SQLExecDirect(m_statement, insert_req, SQL_NTS);
