+++ modules/platforms/cpp/tests/odbc-test/odbc_suite.h
@@ -22,6 +22,7 @@
 #endif
 
 #include "ignite_runner.h"
+#include "odbc_connection.h"
 #include "odbc_test_utils.h"
 #include "test_utils.h"
 
@@ -40,7 +41,7 @@ using namespace std::string_view_literals;
 /**
  * Test suite.
  */
-class odbc_suite : public ::testing::Test {
+class odbc_suite : public ::testing::Test, public odbc_connection {
 public:
     static constexpr std::string_view TABLE_1 = "tbl1"sv;
     static constexpr std::string_view TABLE_NAME_ALL_COLUMNS = "tbl_all_columns"sv;
@@ -72,74 +73,6 @@ public:
     static std::string get_basic_connection_string() {
         return "driver={" + DRIVER_NAME + "};address=" + get_nodes_address() + ';';
     }
-
-    /**
-     * Prepare handles for connection.
-     */
-    void prepare_environment() {
-        ignite::prepare_environment(m_env, m_conn);
-    }
-
-    /**
-     * ODBC connect.
-     *
-     * @param connect_str Connect string.
-     */
-    void odbc_connect(std::string_view connect_str) {
-        ignite::odbc_connect(connect_str, m_env, m_conn, m_statement);
-    }
-
-    /**
-     * Disconnect.
-     */
-    void odbc_disconnect() {
-        ignite::odbc_disconnect(m_conn, m_statement);
-    }
-
-    /**
-     * Clean up handles.
-     */
-    void odbc_clean_up() {
-        ignite::odbc_clean_up(m_env, m_conn, m_statement);
-    }
-
-    /**
-     * Execute query.
-     *
-     * @param qry Query.
-     * @return Result.
-     */
-    SQLRETURN exec_query(const std::string& qry) { // NOLINT(readability-make-member-function-const)
-        auto sql = make_odbc_string(qry);
-        return SQLExecDirect(m_statement, sql.data(), static_cast<SQLINTEGER>(sql.size()));
-    }
-
-    /**
-     * Get statement error state.
-     *
-     * @return Statement error state.
-     */
-    [[nodiscard]] std::string get_statement_error_state() const {
-        return get_odbc_error_state(SQL_HANDLE_STMT, m_statement);
-    }
-
-    /**
-     * Get connection error state.
-     *
-     * @return Connection error state.
-     */
-    [[nodiscard]] std::string get_connection_error_state() const {
-        return get_odbc_error_state(SQL_HANDLE_DBC, m_conn);
-    }
-
-    /** Environment handle. */
-    SQLHENV m_env{SQL_NULL_HANDLE};
-
-    /** Connection handle. */
-    SQLHDBC m_conn{SQL_NULL_HANDLE};
-
-    /** Statement handle. */
-    SQLHSTMT m_statement{SQL_NULL_HANDLE};
 };
 
 } // namespace ignite
